+++ RKReader/Views/RKBottomStatusBar.m
@@ -111,7 +111,7 @@
 - (void)setBook:(RKHomeListBooks *)book {
 	_book = book;
 
-	self.bookName.text = [NSString stringWithFormat:@"%@(%ld/%ld)",self.book.fileInfo.fileName,self.book.readProgress.page+1,self.chapter.pageCount];
+	self.bookName.text = [NSString stringWithFormat:@"%@(%ld/%ld)",self.book.fileInfo.fileName,(long)self.book.readProgress.page+1,(long)self.chapter.pageCount];
 	if (self.chapters == 1) {
 		self.progress.text = [NSString stringWithFormat:@"%.2f%%",self.book.readProgress.page*1.0f/self.chapter.pageCount];
 	}else if (self.chapters == 0) {
